pyproject.toml
[tool.poetry]
name = "pr-review-cli"
version = "0.1.0"
description = "AI-powered PR review assistant for Bitbucket"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
pydantic = "^2.0"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
textual = "^0.86.0"
typer = "^0.12.0"
python-frontmatter = "^1.0.0"

[tool.poetry.scripts]
pr-review = "pr_review.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
